refactor: 스웨거 수정[#125]
